refactor(products): move partials into components and product-icon to features
delete-confirmation and product-design move into products/components. product-icon
is shared and domain-aware, so it moves to features/ and the cross-module import
from orders/details points there. Refs #131.
