+++ resources/js/features/product-icon.tsx
