+++ resources/js/pages/products/components/delete-confirmation.tsx
