+++ resources/js/pages/products/components/product-design.tsx
